refactor: button slider tabs
