+++ src/slider/demo/BasicSlider.tsx
@@ -1,7 +1,25 @@
-import { Slider } from 'kool-ui'
-import React from 'react'
+import { Slider, Switch } from 'kool-ui'
+import React, { useState } from 'react'
 
 const BasicSlider: React.FC = () => {
-  return <Slider range={false} min={10} max={20} defaultValue={15} />
+  const [disable, setDisable] = useState<boolean>(false)
+  return (
+    <div>
+      <Slider
+        vertical={true}
+        range={false}
+        min={10}
+        max={20}
+        defaultValue={15}
+        disable={disable}
+      />
+      <Switch
+        checked={disable}
+        onChange={() => {
+          setDisable(!disable)
+        }}
+      />
+    </div>
+  )
 }
 export default BasicSlider
