+++ src/slider/style/index.scss
@@ -54,8 +54,8 @@
 .koo-slider-short-left {
   height: 10px;
   width: 10px;
-  border-bottom-right-radius: 1rem;
-  border-top-right-radius: 1rem;
+  border-bottom-left-radius: 1rem;
+  border-top-left-radius: 1rem;
   position: absolute;
   display: inline-block;
   background-color: var(--koo-gray-5);
@@ -72,8 +72,8 @@
 
 .koo-slider-short-right {
   height: 10px;
-  border-bottom-left-radius: 1rem;
-  border-top-left-radius: 1rem;
+  border-bottom-right-radius: 1rem;
+  border-top-right-radius: 1rem;
   position: absolute;
   display: inline-block;
   background-color: var(--koo-gray-5);
